refactor: password 타입의 input에서 자동 완성 기능 끄기
App의 submit 함수를 하위 페이지로 이동
