+++ docs/REFACTOR.md
@@ -13,7 +13,7 @@
 - [x] 다음으로 제출을 했을 때 잘못 입력되어 있을 경우 잘못 입력된 입력창을 포커스해주는 기능 구현
 - [x] 영어 혹은 한글 값으로만 사용자 이름 입력 가능하도록 구현하기
 - [x] 인풋을 입력하면 다음 인풋으로 포커스하는 기능 구현하기
-- [ ] input의 password 타입을 사용하지 않고 text로 고쳐서 암호화해서 보여주도록 하기. (password 타입을 사용하면 패스워드 자동 기억 기능이 나타나는 것이 사용자에게 안 좋음)
+- [x] password 타입의 input에서 자동 완성 기능 끄기
 - [ ] 달은 1 ~ 12까지 , 연도는 현재 연도 ~ 현재 연도 +5 까지 입력값 받기 (유효성 검사)
 - [ ] 잘못 입렷한 입력창은 테두리를 빨갛게 보여주는 기능 구현
 - [ ] 에러 메세지 툴팁을 더 고도화 해보기
